engine: Port additional image cleanup integ tests
