+++ misc/image-cleanup-test-images/windows5.dockerfile
@@ -15,8 +15,7 @@ FROM microsoft/windowsservercore:latest
 MAINTAINER Amazon Web Services, Inc.
 
 SHELL ["powershell", "-command"]
-RUN echo 6 > image.txt
-RUN date > file.txt
-
 ENTRYPOINT ["powershell","-command"]
-CMD ["sleep", "30"]
+CMD ["sleep", "5"]
+
+RUN echo 6 > image.txt
